+++ changelog.md
@@ -1,3 +1,4 @@
 # Unreleased
 
 - Change super.onClick on each buttons
+- Add inventory name for commands arguments (https://github.com/Maxlego08/zMenu/issues/12)